Moved matching modals out of applicant-view directory
